renamed QA folder to questions_answers to match group repo
