+++ client/components/questions_answers/helpers.js
